Refactored projects scanning into function

+++ imageconv.py
@@ -5,7 +5,6 @@ from json import load
 
 
 def image_converter(path, out_dir):
-    parent_dir = path.parent
     in_img = Image.open(path)
 
     lum = in_img.convert('L')
@@ -16,6 +15,25 @@ def image_converter(path, out_dir):
 
     in_img.save(Path.joinpath(out_dir, 'thumbnail.jpg'), quality=70, progressive=True)
     out_img.save(Path.joinpath(out_dir, 'thumbnail_blue.jpg'), quality=70, progressive=True)
+
+
+def scan_projects(in_projects_dir, out_projects_dir):
+
+    for in_proj_dir in in_projects_dir.iterdir():
+
+        proj_id = in_proj_dir.stem
+        print(proj_id)
+
+        out_proj_dir = Path.joinpath(out_projects_dir, proj_id)
+        out_proj_dir.mkdir()
+
+        json_fp = open(str(Path.joinpath(in_proj_dir, 'info.json')), 'r')
+        info_json = load(json_fp)
+        projects[proj_id] = info_json
+        json_fp.close()
+
+        # Create thumbnails
+        image_converter(Path.joinpath(in_proj_dir, 'preview.jpg'), out_proj_dir)
 
 
 def build_menu(page):
@@ -49,38 +67,23 @@ def build_project_list(projects, filter = None):
     return projs
 
 
-
 out_dir = Path.joinpath(Path.cwd(), 'out')
-rmtree(str(out_dir))
+if out_dir.is_dir():
+    rmtree(str(out_dir))
 out_dir.mkdir()
 
 in_dir = Path.joinpath(Path.cwd(), 'in')
 
-out_projects = Path.joinpath(out_dir, 'projects')
-out_projects.mkdir()
+out_projects_dir = Path.joinpath(out_dir, 'projects')
+out_projects_dir.mkdir()
 
 in_projects_dir = Path.joinpath(in_dir, 'projects')
 
 projects = {}   # Stores info of all projects
 
-for in_proj_dir in in_projects_dir.iterdir():
-
-    proj_id = in_proj_dir.stem
-    print(proj_id)
-
-    out_proj_dir = Path.joinpath(out_projects, proj_id)
-    out_proj_dir.mkdir()
-
-    json_fp = open(str(Path.joinpath(in_proj_dir, 'info.json')), 'r')
-    info_json = load(json_fp)
-    projects[proj_id] = info_json
-    json_fp.close()
-
-    # Create thumbnails
-    image_converter(Path.joinpath(in_proj_dir, 'preview.jpg'), out_proj_dir)
+scan_projects(in_projects_dir, out_projects_dir)
     
-
-home_html = '''
+out_home_html = '''
 <!DOCTYPE html>
 <html lang="en">
     <head>
@@ -97,8 +100,8 @@ home_html = '''
     </body>
 </html>
 '''.format(menu=build_menu('home'), list=build_project_list(projects))
-home_path = Path.joinpath(out_dir, 'home.html')
-home_path.touch()
-home_fp = open(str(home_path), 'w')
-home_fp.write(home_html)
-home_fp.close()
+out_home_path = Path.joinpath(out_dir, 'home.html')
+out_home_path.touch()
+out_home_fp = open(str(out_home_path), 'w')
+out_home_fp.write(out_home_html)
+out_home_fp.close()
